update PROVISIONAL executable
Put provisional executable code based on the previous commit in /bin_PROVISIONAL. It is provided temporarily for evaluation purposes as a courtesy to users who are unable to compile the source code. Only a 64-bit Windows executable is provided. Please see the disclaimer in the accompanying README_provisional.txt file.